Fix update on different song in album
Currently Google Play Music's Total and Current Time are slightly broken and won't always report back with the correct amount!
(Issue)[https://github.com/MarshallOfSound/Google-Play-Music-Desktop-Player-UNOFFICIAL-/issues/2630]

Because of this, in the If statement on line 174 it won't always trigger if someone is constantly listening through an album. The Album Art stays the same, Google caches the next song so it doesn't pause briefly, and the total time doesn't change due to this bug!